Em-82 prettier switch case
